Resolves: MTV-3884 | Rename OVA proxy to Forklift proxy
