+++ operator/config/manifests/bases/forklift-operator.clusterserviceversion.yaml
@@ -171,9 +171,9 @@ spec:
         path: must_gather_image_fqin
         x-descriptors:
         - urn:alm:descriptor:com.tectonic.ui:hidden
-      - description: OVA inventory proxy image
-        displayName: OVA Proxy Image
-        path: ova_proxy_fqin
+      - description: Forklift inventory proxy image
+        displayName: Forklift Proxy Image
+        path: forklift_proxy_fqin
         x-descriptors:
         - urn:alm:descriptor:com.tectonic.ui:hidden
       # Controller Resource Configuration
@@ -499,30 +499,30 @@ spec:
         path: ova_container_requests_memory
         x-descriptors:
         - urn:alm:descriptor:com.tectonic.ui:hidden
-      # OVA Proxy Settings
-      - description: OVA Proxy CPU limit (default 1000m)
-        displayName: OVA Proxy CPU Limit
-        path: ova_proxy_container_limits_cpu
+      # Forklift Proxy Settings
+      - description: Forklift Proxy CPU limit (default 1000m)
+        displayName: Forklift Proxy CPU Limit
+        path: forklift_proxy_container_limits_cpu
         x-descriptors:
         - urn:alm:descriptor:com.tectonic.ui:hidden
-      - description: OVA Proxy memory limit (default 1Gi)
-        displayName: OVA Proxy Memory Limit
-        path: ova_proxy_container_limits_memory
+      - description: Forklift Proxy memory limit (default 1Gi)
+        displayName: Forklift Proxy Memory Limit
+        path: forklift_proxy_container_limits_memory
         x-descriptors:
         - urn:alm:descriptor:com.tectonic.ui:hidden
-      - description: OVA Proxy CPU request (default 250m)
-        displayName: OVA Proxy CPU Request
-        path: ova_proxy_container_requests_cpu
+      - description: Forklift Proxy CPU request (default 250m)
+        displayName: Forklift Proxy CPU Request
+        path: forklift_proxy_container_requests_cpu
         x-descriptors:
         - urn:alm:descriptor:com.tectonic.ui:hidden
-      - description: OVA Proxy memory request (default 512Mi)
-        displayName: OVA Proxy Memory Request
-        path: ova_proxy_container_requests_memory
+      - description: Forklift Proxy memory request (default 512Mi)
+        displayName: Forklift Proxy Memory Request
+        path: forklift_proxy_container_requests_memory
         x-descriptors:
         - urn:alm:descriptor:com.tectonic.ui:hidden
-      - description: OVA Proxy route timeout (default 360s)
-        displayName: OVA Proxy Route Timeout
-        path: ova_proxy_route_timeout
+      - description: Forklift Proxy route timeout (default 360s)
+        displayName: Forklift Proxy Route Timeout
+        path: forklift_proxy_route_timeout
         x-descriptors:
         - urn:alm:descriptor:com.tectonic.ui:hidden
       # Logging & General Settings
