+++ operator/related_images.yaml
@@ -21,8 +21,8 @@ relatedImages:
     image: ${UI_PLUGIN_IMAGE}
   - name: ova_provider_server
     image: ${OVA_PROVIDER_SERVER_IMAGE}
-  - name: ova_proxy
-    image: ${OVA_PROXY_IMAGE}
+  - name: forklift_proxy
+    image: ${FORKLIFT_PROXY_IMAGE}
   - name: vsphere_xcopy_volume_populator
     image: ${VSPHERE_XCOPY_VOLUME_POPULATOR_IMAGE}
   - name: cli_download
