+++ operator/roles/forkliftcontroller/defaults/main.yml
@@ -158,19 +158,19 @@ ova_container_limits_memory: "1Gi"
 ova_container_requests_cpu: "100m"
 ova_container_requests_memory: "512Mi"
 
-ova_proxy_fqin: "{{ lookup( 'env', 'OVA_PROXY_IMAGE') or lookup( 'env', 'RELATED_IMAGE_OVA_PROXY') }}"
-ova_proxy_container_name: "{{ app_name }}-ova-proxy"
-ova_proxy_service_name: "{{ app_name }}-ova-proxy"
-ova_proxy_route_name: "{{ ova_proxy_service_name }}"
-ova_proxy_route_timeout: "360s"
-ova_proxy_subapp_name: "ova-proxy"
-ova_proxy_container_limits_cpu: "1000m"
-ova_proxy_container_limits_memory: "1Gi"
-ova_proxy_container_requests_cpu: "250m"
-ova_proxy_container_requests_memory: "512Mi"
-ova_proxy_tls_secret_name: "{{ ova_proxy_service_name }}-serving-cert"
-ova_proxy_issuer_name: "{{ ova_proxy_service_name }}-issuer"
-ova_proxy_certificate_name: "{{ ova_proxy_service_name }}-certificate"
+forklift_proxy_fqin: "{{ lookup( 'env', 'FORKLIFT_PROXY_IMAGE') or lookup( 'env', 'RELATED_IMAGE_FORKLIFT_PROXY') }}"
+forklift_proxy_container_name: "{{ app_name }}-forklift-proxy"
+forklift_proxy_service_name: "{{ app_name }}-forklift-proxy"
+forklift_proxy_route_name: "{{ forklift_proxy_service_name }}"
+forklift_proxy_route_timeout: "360s"
+forklift_proxy_subapp_name: "forklift-proxy"
+forklift_proxy_container_limits_cpu: "1000m"
+forklift_proxy_container_limits_memory: "1Gi"
+forklift_proxy_container_requests_cpu: "250m"
+forklift_proxy_container_requests_memory: "512Mi"
+forklift_proxy_tls_secret_name: "{{ forklift_proxy_service_name }}-serving-cert"
+forklift_proxy_issuer_name: "{{ forklift_proxy_service_name }}-issuer"
+forklift_proxy_certificate_name: "{{ forklift_proxy_service_name }}-certificate"
 
 metric_service_name: "{{ app_name }}-metrics"
 metric_servicemonitor_name: "{{ app_name }}-metrics"
