+++ operator/roles/forkliftcontroller/templates/forklift-proxy/ca.yml.j2
@@ -1,0 +1,20 @@
+---
+apiVersion: cert-manager.io/v1
+kind: Certificate
+metadata:
+  name: {{ forklift_proxy_certificate_name }}
+  namespace: {{ app_namespace }}
+spec:
+  isCA: true
+  dnsNames:
+  - {{ forklift_proxy_service_name }}.{{ app_namespace }}.svc
+  - {{ forklift_proxy_service_name }}.{{ app_namespace }}.svc.cluster.local
+  commonName: {{ forklift_proxy_certificate_name }}
+  secretName: {{ forklift_proxy_tls_secret_name }}
+  privateKey:
+    algorithm: ECDSA
+    size: 256
+  issuerRef:
+    name: {{ app_name }}-issuer
+    kind: Issuer
+    group: cert-manager.io
