+++ operator/roles/forkliftcontroller/templates/forklift-proxy/service.yml.j2
@@ -4,13 +4,13 @@ kind: Service
 metadata:
 {% if not k8s_cluster|bool %}
   annotations:
-    service.beta.openshift.io/serving-cert-secret-name: {{ ova_proxy_tls_secret_name }}
+    service.beta.openshift.io/serving-cert-secret-name: {{ forklift_proxy_tls_secret_name }}
 {% endif %}
   labels:
     app: {{ app_name }}
-    subapp: {{ ova_proxy_subapp_name }}
-    service: {{ ova_proxy_service_name }}
-  name: {{ ova_proxy_service_name }}
+    subapp: {{ forklift_proxy_subapp_name }}
+    service: {{ forklift_proxy_service_name }}
+  name: {{ forklift_proxy_service_name }}
   namespace: {{ app_namespace }}
 spec:
   ports:
@@ -20,4 +20,4 @@ spec:
     protocol: TCP
   selector:
     app: {{ app_name }}
-    subapp: {{ ova_proxy_subapp_name }}
+    subapp: {{ forklift_proxy_subapp_name }}
