--- operator/roles/forkliftcontroller/templates/ova-proxy/ca.yml.j2
@@ -1,20 +1,0 @@
----
-apiVersion: cert-manager.io/v1
-kind: Certificate
-metadata:
-  name: {{ ova_proxy_certificate_name }}
-  namespace: {{ app_namespace }}
-spec:
-  isCA: true
-  dnsNames:
-  - {{ ova_proxy_service_name }}.{{ app_namespace }}.svc
-  - {{ ova_proxy_service_name }}.{{ app_namespace }}.svc.cluster.local
-  commonName: {{ ova_proxy_certificate_name }}
-  secretName: {{ ova_proxy_tls_secret_name }}
-  privateKey:
-    algorithm: ECDSA
-    size: 256
-  issuerRef:
-    name: {{ app_name }}-issuer
-    kind: Issuer
-    group: cert-manager.io
