--- operator/roles/forkliftcontroller/templates/ova-proxy/deployment.yml.j2
@@ -1,58 +1,0 @@
----
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  labels:
-    app: {{ app_name }}
-    subapp: {{ ova_proxy_subapp_name }}
-  name: {{ ova_proxy_service_name }}
-  namespace: {{ app_namespace }}
-spec:
-  selector:
-    matchLabels:
-      app: {{ app_name }}
-      subapp: {{ ova_proxy_subapp_name }}
-  template:
-    metadata:
-      labels:
-        app: {{ app_name }}
-        subapp: {{ ova_proxy_subapp_name }}
-    spec:
-      serviceAccountName: {{ controller_service_name }}
-      containers:
-      - name: main
-        command:
-        - /usr/local/bin/ova-proxy
-        env:
-        - name: POD_NAMESPACE
-          valueFrom:
-            fieldRef:
-              fieldPath: metadata.namespace
-        - name: PORT
-          value: "8443"
-        - name: TLS_CERTIFICATE
-          value: "/var/run/secrets/{{ ova_proxy_tls_secret_name }}/tls.crt"
-        - name: TLS_KEY
-          value: /var/run/secrets/{{ ova_proxy_tls_secret_name }}/tls.key
-        image: {{ ova_proxy_fqin }}
-        imagePullPolicy: {{ image_pull_policy }}
-        resources:
-          limits:
-            cpu: {{ ova_proxy_container_limits_cpu }}
-            memory: {{ ova_proxy_container_limits_memory }}
-          requests:
-            cpu: {{ ova_proxy_container_requests_cpu }}
-            memory: {{ ova_proxy_container_requests_memory }}
-        ports:
-        - name: api
-          containerPort: 8443
-          protocol: TCP
-        volumeMounts:
-        - mountPath: /var/run/secrets/{{ ova_proxy_tls_secret_name }}
-          name: {{ ova_proxy_tls_secret_name }}
-      terminationGracePeriodSeconds: 10
-      volumes:
-      - name: {{ ova_proxy_tls_secret_name }}
-        secret:
-          defaultMode: 420
-          secretName: {{ ova_proxy_tls_secret_name }}
