--- operator/roles/forkliftcontroller/templates/ova-proxy/route.yml.j2
@@ -1,19 +1,0 @@
----
-apiVersion: route.openshift.io/v1
-kind: Route
-metadata:
-  name: {{ ova_proxy_route_name }}
-  namespace: {{ app_namespace }}
-  annotations:
-    haproxy.router.openshift.io/timeout: {{ ova_proxy_route_timeout | default('360s', true) }}
-  labels:
-    app: {{ app_name }}
-    subapp: {{ ova_proxy_subapp_name }}
-    service: {{ ova_proxy_service_name }}
-spec:
-  to:
-    kind: Service
-    name: {{ ova_proxy_service_name }}
-  tls:
-    termination: reencrypt
-    insecureEdgeTerminationPolicy: Redirect
